Enable parallel tests and caching on CI (#406)
* cache `_build`, `deps` and dialyzer `.*plts` on CI
* split tests on CI
