+++ .gitignore
@@ -1,4 +1,5 @@
 .elixir_ls
+priv/plts
 
 # Created by https://www.gitignore.io/api/c,vim,linux,macos,elixir,windows,visualstudiocode
