+++ config/config.exs
@@ -1,0 +1,4 @@
+import Config
+
+config :junit_formatter,
+  include_filename?: true
