+++ test/test_helper.exs
@@ -1,1 +1,2 @@
+ExUnit.configure(formatters: [ExUnit.CLIFormatter, JUnitFormatter])
 ExUnit.start(exclude: [:long_running], capture_log: true, assert_receive_timeout: 500)
